pyproject.toml
[project]
name = "lpilGerbyTagTools"
version = "0.0.1"
description = "A collection of tools to manage LPiL Gerby tag databases."
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "lpilGerbyConfig @ git+https://github.com/litProgLaTeX/lpil-gerby-configuration.git",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
lgtExporter  = "lgtTools.exporter:cli"
lgtImporter  = "lgtTools.importer:cli"
lgtScanner   = "lgtTools.scanner:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
